Restructure; Tests pass /w async-asgi-testclient
