+++ .circleci/environment.yml
@@ -8,5 +8,9 @@ dependencies:
   - jinja2
   - bs4
   - pytest
+  - pytest-asyncio
   - pytest-cov
   - codecov
+  - pip
+  - pip:
+    - async-asgi-testclient
