+++ setup.cfg
@@ -6,4 +6,4 @@ max-complexity = 20
 [isort]
 line_length = 100
 known_first_party = routes,views
-known_third_party = aiohttp,aiohttp_jinja2,bs4,click,fastapi,jinja2,mockupdb,motor,pytest
+known_third_party = aiohttp,aiohttp_jinja2,async_asgi_testclient,bs4,click,fastapi,jinja2,mockupdb,motor,pytest
